==> 1.combining train-test and get the labels.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt 
import datetime

data_path = '../../kaggleData/JD_logging/'
feature_path = '../../kaggleData/JD_logging/features/'

login_train = pd.read_csv(data_path+'t_login.csv')
trade_train = pd.read_csv(data_path+'t_trade.csv')
login_test = pd.read_csv(data_path+'t_login_test.csv')
trade_test = pd.read_csv(data_path+'t_trade_test.csv')

class Config:
    pass
config = Config()

config.trade_train_size = trade_train.shape[0]

config.trade_train_size

login_train['from'] = 'train'
login_test['from'] = 'test'
trade_train['from'] = 'train'
trade_test['from'] = 'test'

login_tt = pd.concat([login_train,login_test])
trade_tt = pd.concat([trade_train.drop('is_risk',axis = 1),trade_test])
trade_tt['time'] = trade_tt['time'].apply(lambda x : x[0:19])

trade_label = trade_train['is_risk']

#testing recombining
trade_new_train = trade_tt.iloc[:config.trade_train_size,:]
trade_new_train['is_risk'] = trade_label

print trade_train.shape[0]
print pd.concat([trade_new_train,trade_train]).drop_duplicates().shape[0]

trade_tt.to_csv(data_path+'trade_tt.csv',index = False)
login_tt.to_csv(data_path+'login_tt.csv',index = False)
trade_label.to_pickle(data_path+'trade_train_lable.pkl')

import pickle
output = open('config.pkl', 'wb')
pickle.dump(config,output)

trade_tt[config.trade_train_size:]['rowkey']

pd.to_pickle(trade_tt[config.trade_train_size:]['rowkey'],data_path+'trade_test_rowkey.pkl')
