==> 1.5 Setting the validation set and train set indice.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt 
import datetime

class Config:
    pass
config = pd.read_pickle('config.pkl')
data_path = config.data_path
feature_path = config.feature_path
print dir(config)

trade_tt = pd.read_csv(data_path+'trade_tt_with_recent_3_log_id.csv')

trade_tt.columns

#set the seed
np.random.seed(314)
def get_certain_part(sequence,numerator,denominator):
    return sequence[(len(sequence)*(numerator-1)/denominator):(len(sequence)*numerator/denominator)]

#get the indice in the certain month and shuffle
trade_tt_2_4_index = list(trade_tt[(trade_tt['month']>1)&(trade_tt['month']<5)].index)
trade_tt_3_5_index = list(trade_tt[(trade_tt['month']>2)&(trade_tt['month']<6)].index)
trade_tt_5_index = list(trade_tt[trade_tt['month']==5].index)
trade_tt_6_index =list(trade_tt[trade_tt['month']==6].index)

np.random.shuffle(trade_tt_2_4_index)
np.random.shuffle(trade_tt_3_5_index)
np.random.shuffle(trade_tt_5_index)
np.random.shuffle(trade_tt_6_index)

#randomly split the indice into 3 parts and making tuple for each train/validation set
validation_indice_tuple_list = []

for i in range(3):
    validation_indice_tuple_list.append((get_certain_part(trade_tt_2_4_index,i+1,3),get_certain_part(trade_tt_5_index,i+1,3)))
    validation_indice_tuple_list.append((get_certain_part(trade_tt_3_5_index,i+1,3),get_certain_part(trade_tt_6_index,i+1,3)))
print len(validation_indice_tuple_list)

trade_tt_2_6_index = list(trade_tt[(trade_tt['month']>1)&(trade_tt['month']<7)].index)
config.train_2_6_index = trade_tt_2_6_index
config.single_module_validation_indice_set = validation_indice_tuple_list
pd.to_pickle(config,'config.pkl')
