==> 2.2 addtional feature for login details.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt 
import datetime

class Config:
    pass
config = pd.read_pickle('config.pkl')
data_path = '../../kaggleData/JD_logging/'
feature_path = '../../kaggleData/JD_logging/features/'

config.feature_dict

login_tt = pd.read_csv(data_path+'login_tt.csv')
#trade_tt = pd.read_csv(data_path+'trade_tt.csv')
login_tt['time'] = login_tt['time'].apply(lambda x : datetime.datetime.strptime(x , '%Y-%m-%d %H:%M:%S'))
#trade_tt['time'] = trade_tt['time'].apply(lambda x : datetime.datetime.strptime(x , '%Y-%m-%d %H:%M:%S'))

#login_tt['time'] = login_tt['time'].apply(lambda x : datetime.datetime.strptime(x , '%Y-%m-%d %H:%M:%S'))

#login time generating and building cycles
login_tt['month'] = login_tt['time'].apply(lambda x : x.month)
login_tt['day'] = login_tt['time'].apply(lambda x : x.day)
login_tt['weekday'] = login_tt['time'].apply(lambda x : x.weekday())
login_tt['hour'] = login_tt['time'].apply(lambda x : x.hour)

#login_tt['month_cycle'] = login_tt['month'].apply(lambda x : x + 6 if x + 6 < 12 else x-6)
login_tt['day_cycle'] = login_tt['day'].apply(lambda x : x+15 if x + 15< 31 else x - 15)
login_tt['weekday_cycle'] = login_tt['weekday'].apply(lambda x : x+3 if x +3 < 7 else x -3)
login_tt['hour_cycle'] = login_tt['hour'].apply(lambda x : x + 12 if x + 12 < 23 else x - 12)

"""
#trade time generating and building cycles
trade_tt['month'] = trade_tt['time'].apply(lambda x : x.month)
trade_tt['day'] = trade_tt['time'].apply(lambda x : x.day)
trade_tt['weekday'] = trade_tt['time'].apply(lambda x : x.weekday())
trade_tt['hour'] = trade_tt['time'].apply(lambda x : x.hour)

#trade_tt['month_cycle'] = trade_tt['month'].apply(lambda x : x + 6 if x + 6 < 12 else x-6)
trade_tt['day_cycle'] = trade_tt['day'].apply(lambda x : x+15 if x + 15< 31 else x - 15)
trade_tt['weekday_cycle'] = trade_tt['weekday'].apply(lambda x : x+3 if x +3 < 7 else x -3)
trade_tt['hour_cycle'] = trade_tt['hour'].apply(lambda x : x + 12 if x + 12 < 23 else x - 12)

"""

#trade_tt.to_csv(data_path+'trade_tt_with_recent_3_log_id.csv',index =False)
trade_tt = pd.read_csv(data_path+'trade_tt_with_recent_3_log_id.csv')
trade_tt['time'] = trade_tt['time'].apply(lambda x : datetime.datetime.strptime(x , '%Y-%m-%d %H:%M:%S'))

detail_info = ['timelong','log_from','city','result','type','is_scan',
               'month','day','weekday','hour','day_cycle','weekday_cycle','hour_cycle']
login_tt_detail = login_tt[['log_id','time','device','ip']+detail_info]

#merging
trade_tt = trade_tt.merge(login_tt_detail.add_suffix('_login_0'),how = 'left',left_on ='recent_login_id_0',right_on = 'log_id_login_0',
                          copy = False)
trade_tt = trade_tt.merge(login_tt_detail.add_suffix('_login_1'),how = 'left',left_on ='recent_login_id_1',right_on = 'log_id_login_1',
                         copy = False)
trade_tt = trade_tt.merge(login_tt_detail.add_suffix('_login_2'),how = 'left',left_on ='recent_login_id_2',right_on = 'log_id_login_2',
                          copy = False)

#getting the time distance 
trade_tt['login_distance_0']=trade_tt['time']-trade_tt['time_login_0']
trade_tt['login_distance_1']=trade_tt['time']-trade_tt['time_login_1']
trade_tt['login_distance_2']=trade_tt['time']-trade_tt['time_login_2']

trade_tt['login_distance_0'] = trade_tt['login_distance_0'].apply(lambda x : x.total_seconds())
trade_tt['login_distance_1'] = trade_tt['login_distance_1'].apply(lambda x : x.total_seconds())
trade_tt['login_distance_2'] = trade_tt['login_distance_2'].apply(lambda x : x.total_seconds())

trade_tt.columns

### filling the login_distance_0/1/2

Use following rules:
   start_date datetime.datetime.strptime('2015-1-1','%Y-%m-%d')
   - mean_login_distance, last login distance from start if no login, single login time from start or to final day if only one case
   - for login_distance 0, return max(mean_login_distance,row distance from start_date)
   - for login distance 1, return max(mean_login_distance+login_distance 0，row distance from start_date)
   - if login_distance 0 don't exist then return mean_login_distance+login_distance 0 for login distance 1

"""
works on spark
"""

import pyspark as spark
date_range_list = [360,30,7]

conf = spark.SparkConf().setAppName('jupyter_backend').setMaster('local[15]')\
        .set('spark.executor.memory','2g')\
        .set('spark.default.parallelism','112')
sc = spark.SparkContext(conf=conf)


start_date = datetime.datetime.strptime('2015-1-1','%Y-%m-%d')
end_date = datetime.datetime.strptime('2015-8-1','%Y-%m-%d')

def get_id_mean_login_distance(user_id,login_table):
    related_logins = login_table[login_table['id']==user_id].sort_values(by = 'time')
    if len(related_logins) == 0:
        return (end_date-start_date).total_seconds()
    elif len(related_logins) ==1:
        only_sample_time = related_logins.iloc[0].time
        return min(end_date - only_sample_time,only_sample_time - start_date).total_seconds()
    else:
        delta_list = []
        for i in range(len(related_logins)-1):
            delta_list.append(related_logins.iloc[i+1].time - related_logins.iloc[i].time)
        return np.mean(delta_list).total_seconds()

rdd_buffer = trade_tt['id'].unique()
rdd = sc.parallelize(rdd_buffer)

result_rdd = rdd.map(lambda x : (x,get_id_mean_login_distance(x,login_tt)))
id_result_list = result_rdd.collect()

id_mean_login_distacne  = pd.DataFrame(id_result_list,columns = ['id','mean_login_distance'])
trade_tt = trade_tt.merge(id_mean_login_distacne,how = 'left', on = 'id')

def filling_the_login_distance(idx_row_tuple):
    idx =  idx_row_tuple[0]
    row =  idx_row_tuple[1]
    result_dict = {}
    if np.isnan(row.login_distance_0):
        login_distance_0_al = max(row.mean_login_distance,(row.time - start_date).total_seconds())
        login_distance_1_al = login_distance_0_al + row.mean_login_distance
        login_distance_2_al= login_distance_1_al + row.mean_login_distance
    elif np.isnan(row.login_distance_1):
        login_distance_0_al = row.login_distance_0
        login_distance_1_al = max(row.mean_login_distance+row.login_distance_0,(row.time - start_date).total_seconds())
        login_distance_2_al= login_distance_1_al + row.mean_login_distance
    elif np.isnan(row.login_distance_2):
        login_distance_0_al = row.login_distance_0
        login_distance_1_al = row.login_distance_1
        login_distance_2_al= max(row.mean_login_distance+row.login_distance_1,(row.time - start_date).total_seconds())
    else:
        login_distance_0_al = row.login_distance_0
        login_distance_1_al = row.login_distance_1
        login_distance_2_al= row.login_distance_2
    
    return (idx,row.rowkey,login_distance_0_al,login_distance_1_al,login_distance_2_al)

# begin to filling the login_distance related features
trade_tt_temp = trade_tt[['rowkey','mean_login_distance','login_distance_0',
                         'login_distance_1','login_distance_2','time']]
rdd_buffer = []
for (idx,row) in trade_tt.iterrows():
    rdd_buffer.append((idx,row))
    
rdd = sc.parallelize(rdd_buffer)

result_rdd = rdd.map(lambda x : filling_the_login_distance(x))
new_login_distance_result_dict = result_rdd.collect()

trade_tt_new_login_result = pd.DataFrame(new_login_distance_result_dict,columns = ['ori_id','rowkey','login_distance_0_al',
                                        'login_distance_1_al','login_distance_2_al'])

assert (trade_tt_new_login_result['ori_id']==trade_tt.index).all
assert (trade_tt_new_login_result['rowkey']==trade_tt['rowkey']).all

trade_tt['login_distance_0_al'] = np.log(trade_tt_new_login_result['login_distance_0_al'])
trade_tt['login_distance_1_al'] = np.log(trade_tt_new_login_result['login_distance_1_al'])
trade_tt['login_distance_2_al'] = np.log(trade_tt_new_login_result['login_distance_2_al'])

### timelong % 1000 for whole second check 

### and get the distance of login distance

trade_tt['timelong_login_0_mod_1000'] = trade_tt['timelong_login_0'].apply(lambda x : 1000 - x%1000 if x%1000 >500 else x%1000)
trade_tt['timelong_login_1_mod_1000'] = trade_tt['timelong_login_1'].apply(lambda x : 1000 - x%1000 if x%1000 >500 else x%1000)
trade_tt['timelong_login_2_mod_1000'] = trade_tt['timelong_login_2'].apply(lambda x : 1000 - x%1000 if x%1000 >500 else x%1000)

trade_tt['login_distance_diff']=np.abs(trade_tt['login_distance_2'] + trade_tt['login_distance_0'] - 
                                 2*trade_tt['login_distance_1'])

trade_tt = trade_tt.fillna(-10)

login_detail_new_features = ['login_distance_0_al','login_distance_1_al','login_distance_2_al','mean_login_distance',
                             'timelong_login_0_mod_1000','timelong_login_1_mod_1000','timelong_login_2_mod_1000','login_distance_diff']
config.feature_dict['login_detail_new_features'] = login_detail_new_features
pd.to_pickle(config,'config.pkl')

for feature in login_detail_new_features:
    pd.to_pickle(trade_tt[feature].values,feature_path+feature+'.pkl')
