==> 12 label encoding .ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt 
import datetime

class Config:
    pass
config = pd.read_pickle('config.pkl')

data_path = '../../kaggleData/JD_logging/'
feature_path = '../../kaggleData/JD_logging/features/'
print(dir(config))

login_tt = pd.read_csv(data_path+'login_tt.csv')
trade_tt_with_recent_logins = pd.read_csv(data_path+'trade_tt_with_recent_3_log_id.csv')
trade_tt = pd.read_csv(data_path+'trade_tt.csv')

login_tt['time'] = login_tt['time'].apply(lambda x : datetime.datetime.strptime(x , '%Y-%m-%d %H:%M:%S'))
trade_tt['time'] = trade_tt['time'].apply(lambda x : datetime.datetime.strptime(x , '%Y-%m-%d %H:%M:%S'))
trade_tt_with_recent_logins['time'] = trade_tt_with_recent_logins['time'].apply(lambda x : datetime.datetime.strptime(x , '%Y-%m-%d %H:%M:%S'))

train_labels = pd.read_pickle(data_path+'trade_train_label.pkl')
trade_tt_with_recent_logins.loc[:config.trade_train_size,'is_risk'] = train_labels

del trade_tt_with_recent_logins['month']
del trade_tt_with_recent_logins['day']
del trade_tt_with_recent_logins['weekday']
del trade_tt_with_recent_logins['hour']
del trade_tt_with_recent_logins['day_cycle']
del trade_tt_with_recent_logins['weekday_cycle']
del trade_tt_with_recent_logins['hour_cycle']
del trade_tt_with_recent_logins['recent_login_id_1']
del trade_tt_with_recent_logins['recent_login_id_2']
trade_wit_last_login_tt = trade_tt_with_recent_logins.merge(login_tt, how = 'left',left_on ='recent_login_id_0',right_on = 'log_id',suffixes=('', '_login'))

trade_wit_last_login_tt

test_row = trade_wit_last_login_tt.iloc[45050]

def find_related_hcc_trades_before(row,trade_table,hcc,*args,**kw):
    related_trades = trade_table[trade_table[hcc] == row[hcc]]
    related_trades_before = related_trades[related_trades['time']<row.time]
    return related_trades_before    

def get_mult_hcc_target_encoding(idx_row_tuple,trade_table,hcc_list):
    idx = idx_row_tuple[0]
    row = idx_row_tuple[1]
    result_dict= {}
        
    for feature in hcc_list:
        result_dict[feature+'_t_encoding'] = np.mean(find_related_hcc_trades_before(row,trade_table,feature)['is_risk'].dropna())
    return (idx,row.rowkey,result_dict)

get_mult_hcc_target_encoding((45050,test_row),trade_wit_last_login_tt,['id','ip','device','city'])

### make new hcc features by combining

trade_wit_last_login_tt['ip_id']  = trade_wit_last_login_tt.apply(lambda x : str((x.ip,x.id)),axis = 1)
trade_wit_last_login_tt['ip_device']  = trade_wit_last_login_tt.apply(lambda x : str((x.ip,x.device)),axis = 1)
trade_wit_last_login_tt['city_id']  = trade_wit_last_login_tt.apply(lambda x : str((x.city,x.id)),axis = 1)
trade_wit_last_login_tt['city_device']  = trade_wit_last_login_tt.apply(lambda x : str((x.city,x.device)),axis = 1)
trade_wit_last_login_tt['device_id']  = trade_wit_last_login_tt.apply(lambda x : str((x.id,x.device)),axis = 1)


print(trade_wit_last_login_tt.shape)
print(len(trade_wit_last_login_tt['ip_id'].unique()))
print(len(trade_wit_last_login_tt['ip_device'].unique()))
print(len(trade_wit_last_login_tt['city_id'].unique()))
print(len(trade_wit_last_login_tt['city_device'].unique()))
print(len(trade_wit_last_login_tt['device_id'].unique()))

trade_wit_last_login_tt['device_id_ip']  = trade_wit_last_login_tt.apply(lambda x : str((x.id,x.device,x.ip)),axis = 1)
trade_wit_last_login_tt['device_id_city']  = trade_wit_last_login_tt.apply(lambda x : str((x.id,x.device,x.city)),axis = 1)
print(len(trade_wit_last_login_tt['device_id_ip'].unique()))
print(len(trade_wit_last_login_tt['device_id_city'].unique()))

### Generating new features with spark

import pyspark as spark

conf = spark.SparkConf().setAppName('jupyter_backend').setMaster('local[12]')\
        .set('spark.executor.memory','2g')\
        .set('spark.default.parallelism','112')
sc = spark.SparkContext(conf=conf)

#parallel computing
hcc_list = ['ip_device','ip_id','device_id','city_id','device_id_city']+['id','ip','device','city']

trade_tt_rdd_buffer = []
for (idx,row) in trade_wit_last_login_tt.iterrows():
    trade_tt_rdd_buffer.append((idx,row))
trade_tt_rdd = sc.parallelize(trade_tt_rdd_buffer)

#result_rdd = trade_tt_rdd.map(lambda idx_row_tuple : get_mult_hcc_target_encoding(idx_row_tuple,trade_wit_last_login_tt,['id','ip','device']))
result_rdd = trade_tt_rdd.map(lambda idx_row_tuple : get_mult_hcc_target_encoding(idx_row_tuple,trade_wit_last_login_tt,hcc_list))

result_rdd_buffer = result_rdd.collect()

result_rdd_to_df_buffer = []
for ori_id,rowkey,result_dict in result_rdd_buffer:
    unit_dict= result_dict
    unit_dict['ori_id'] = ori_id
    unit_dict['rowkey'] = rowkey
    
    result_rdd_to_df_buffer.append(unit_dict)

result_df = pd.DataFrame(result_rdd_to_df_buffer)

assert (np.sum(result_df['rowkey']!=trade_tt['rowkey']))==0
assert (np.sum(result_df['ori_id']!=trade_tt.index))==0

result_df = result_df.fillna(0)

del result_df['rowkey']
del result_df['ori_id']
trade_tt = pd.concat([trade_tt,result_df],axis = 1)
print(trade_tt.shape)

original_hcc_encoding = ['city_t_encoding', 'device_t_encoding', 'id_t_encoding', 'ip_t_encoding']
new_hcc_encodin_double_combining = ['ip_device_t_encoding','ip_id_t_encoding','device_id_t_encoding','city_id_t_encoding','device_id_city_t_encoding']

trade_tt.shape

#get random mask for the encoding features
np.random.seed(242)
random_mask = np.random.uniform(-0.1,0.1,150594)+1

for feature in original_hcc_encoding+new_hcc_encodin_double_combining:
    trade_tt[feature] = trade_tt[feature]*random_mask

for feature in original_hcc_encoding+new_hcc_encodin_double_combining:
    pd.to_pickle(trade_tt[feature].values,feature_path+feature+'.pkl')

config.feature_dict.keys()

config.feature_dict['hcc_target_encoding'] = original_hcc_encoding
config.feature_dict['hcc_mult_target_encoding'] = new_hcc_encodin_double_combining

pd.to_pickle(config,'config.pkl')
