==> 17 data nomalization.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt 
import datetime
from sklearn.metrics import fbeta_score,roc_auc_score
#from mochi import f_beta_01
import re

from scipy.stats import boxcox
from sklearn.preprocessing import normalize

from mochi import check_nan_inf

class Config:
    pass
config = pd.read_pickle('config.pkl')
data_path = config.data_path
feature_path = config.feature_path


normalized_path = '../../kaggleData/JD_logging/normalized/'

features =   (config.feature_dict['trade_detail_feature']+
              config.feature_dict['recent_login_detail']+
              config.feature_dict['trade_and_recent_login_comparing']+
              config.feature_dict['login_trade_hist_stats']+
              config.feature_dict['llc_user_habbit']+
              config.feature_dict['hcc_user_habbit']+
              config.feature_dict['hcc_properties']+
              config.feature_dict['hcc_target_encoding']+
              config.feature_dict['login_detail_new_features']+
              config.feature_dict['hcc_trade_properties']+
              config.feature_dict['hcc_mult_target_encoding']+
              config.feature_dict['hcc_user_trade_habbit']
             )
feature_sequence_list = []
for feature in features:
    feature_sequence_list.append(pd.read_pickle(feature_path+feature+'.pkl').reshape(-1,1))
    
trade_tt_mat = np.hstack(feature_sequence_list)
#trade_tt_mat[trade_tt_mat==-10]=np.nan

validation_tuple_list = config.single_module_validation_indice_set
train_labels = pd.read_pickle(data_path+'trade_train_label.pkl')

### generating the feature set

#checking the filled nan by -10, computing the emptyness
trade_df = pd.DataFrame(trade_tt_mat,columns = features)
trade_df[trade_df ==-10] = np.nan
train_trade = trade_df.iloc[config.train_2_6_index]


#print(dir(config))
print('Currently has feature-sets')
print(config.feature_dict.keys())

processed_features = []
#processed_feature_names = []

### Check data type for features

types = trade_df['device_comparing_login_1'].apply(lambda x : type(x)).unique()

type_list = []
type_set =set([])
for feature in features:
    unique_list = list(trade_df[feature].apply(lambda x : type(x)).unique())
    type_list.append((feature,unique_list))
    type_set.update(unique_list)

### dealing with boolean features

#boolean features
boolean_features = []
for feature,unique_types in type_list:
    if bool in unique_types:
        boolean_features.append((feature,unique_types))
boolean_features

#filling the -10 to be false
feature_with_nan = [x[0] for x in boolean_features if float in x[1]]

for feature in feature_with_nan:
    trade_df[feature]= trade_df[feature].fillna(False)

processed_features.extend(x[0] for x in boolean_features)

### dealing with other types, check numerical or categorical first

#check numerical or categorical by getting the unique value size
other_features = [x[0] for x in type_list if bool not in x[1]]

other_features_unique_size = [(x,trade_df[x].unique().shape[0]) for x in other_features]
other_features_unique_size = sorted(other_features_unique_size,key = lambda x:x[1],)

other_features_unique_size

- categorical features including :
- type_login_
- log_from_login
- result_login_

categorical_features = []
numerical_features = [x[0] for x in other_features_unique_size]
to_be_remap = []

for feature,size in other_features_unique_size:
    
    if re.match('^type_login_.*',feature) or re.match('^log_from_login.*',feature) or  re.match('^result_login_.*',feature):
        categorical_features.append(feature)
        numerical_features.remove(feature)
    if re.match('.*cycle.*',feature):
        numerical_features.remove(feature)
    elif re.match('^hour.*',feature) or re.match('^weekday.*',feature) or re.match('^day.*',feature):
        to_be_remap.append(feature)
        numerical_features.remove(feature)
        
    

### Apply box_cox on some of the numerical features

def fill_median_and_apply_boxcox_and_nomalize(feature):
    #deal with nan
    if np.sum(np.isnan(trade_df[feature].astype(float))):
        temp_column = trade_df[feature].fillna(train_trade[feature].dropna().median()).astype(float)
    else:
        temp_column = trade_df[feature].astype(float)
        
    #check the minimun
    if temp_column.min()<=0:
        print('meet value smaller than 0 at %s, minimun is %f' % (feature,temp_column.min()))
        temp_column+= (np.abs(temp_column.min())+1)
       
        
    #apply box_cox for the feature 
    result = normalize(boxcox(temp_column)[0].reshape(-1,1))
    
    return result.reshape(-1,)

for feature in numerical_features:
    trade_df[feature] = fill_median_and_apply_boxcox_and_nomalize(feature)
processed_features.extend(numerical_features)

### remaping some of the features to be new categorical

def map_hour(hour):
    if hour ==0 or hour>17:
        #18-0
        return 0
    elif hour>14:
        #15-17
        return 1
    elif hour>11:
        #12-15
        return 2
    elif hour>7:
        #8-11
        return 3
    else:
        #1-7
        return 4

def map_weekday(weekday):
    if weekday ==6 or weekday ==7:
        return True
    else:
        return False

def map_day(day):
    if day < 7:
        return 0
    elif day<15:
        return 1
    elif day<22:
        return 2
    else:
        return 3

for feature in to_be_remap:
    if re.match('^hour.*',feature):
        trade_df[feature] = trade_df[feature].apply(lambda x : map_hour(x))
    elif re.match('^weekday.*',feature):
        trade_df[feature] = trade_df[feature].apply(lambda x : map_weekday(x))
    elif re.match('^day.*',feature):
        trade_df[feature] = trade_df[feature].apply(lambda x : map_day(x))

categorical_features.extend(to_be_remap)

### mapping the categorical features

for feature in categorical_features:
    possible_values = trade_df[feature].unique()
    for value in possible_values:
        trade_df[feature+'_'+str(value)] = (trade_df[feature]==value)
        processed_features.append(feature+'_'+str(value))

### Check and generate the result feature set

trade_tt_mat = trade_df[processed_features]
for feature in [x[0] for x in boolean_features]:
    trade_tt_mat[feature] = trade_tt_mat[feature].apply(lambda x : 1 if x else 0)
check_nan_inf(trade_tt_mat,processed_features)

### generate the normalized features

for feature in processed_features:
    pd.to_pickle(trade_tt_mat[feature].values,normalized_path+feature+'.pkl')

config = pd.read_pickle('config.pkl')
config.normalized_path =normalized_path
config.normalized_features = processed_features
config.numerical_features = numerical_features

pd.to_pickle(config,'config.pkl')
